requirements_lin.txt
arrow==1.3.0
asgiref==3.8.1
async-timeout==4.0.3
click==8.1.7
colorama==0.4.6
diff-match-patch==20230430
Django==5.1.1
django-cors-headers==4.5.0
django-debug-toolbar==4.4.6
django-filter==24.3
django-import-export==4.1.1
django-redis==5.4.0
django-rq==2.10.2
djangorestframework==3.15.0
Markdown==3.7
python-dateutil==2.9.0.post0
redis==5.1.0
rq==1.16.2
six==1.16.0
sqlparse==0.5.1
tablib==3.5.0
times==0.7
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
